+++ source-local/src/commonMain/kotlin/tachiyomi/source/local/io/Format.kt
@@ -8,6 +8,7 @@ sealed interface Format {
     data class Directory(val file: UniFile) : Format
     data class Archive(val file: UniFile) : Format
     data class Epub(val file: UniFile) : Format
+    data class Pdf(val file: UniFile) : Format
 
     class UnknownFormatException : Exception()
 
@@ -16,6 +17,7 @@ sealed interface Format {
         fun valueOf(file: UniFile) = when {
             file.isDirectory -> Directory(file)
             file.extension.equals("epub", true) -> Epub(file)
+            file.extension.equals("pdf", true) -> Pdf(file)
             isArchiveSupported(file) -> Archive(file)
             else -> throw UnknownFormatException()
         }